HQL Reports, Scripts: Delete Confirmation
- added delete confirmation for Scripts and Reports pages

+++ JavaSource/org/unitime/timetable/gwt/client/admin/ScriptPage.java
@@ -738,6 +738,7 @@ public class ScriptPage extends Composite {
 				@Override
 				public void onClick(ClickEvent event) {
 					if (iScriptId != null) {
+						if (!Window.confirm(MESSAGES.confirmDeleteScript())) return;
 						hide();
 						LoadingWidget.getInstance().show(MESSAGES.waitDelete(iName.getText()));
 						RPC.execute(new DeleteScriptRpcRequest(iScriptId, iName.getText()), new AsyncCallback<ScriptInterface>() {
